# src/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.MciamPermission:
    properties:
      action:
        description: e.g., create, read, update, delete
        type: string
      createdAt:
        description: Match DB schema
        type: string
      description:
        type: string
      frameworkId:
        description: FK to mcmp_resource_types.framework_id
        type: string
      id:
        description: 'Format: <framework_id>:<resource_type_id>:<action>'
        type: string
      name:
        type: string
      resourceTypeId:
        description: FK to mcmp_resource_types.id
        type: string
      updatedAt:
        description: Match DB schema
        type: string
    type: object
...
host: localhost
info:
  contact: {}
  description: MC IAM Manager API Documentation
  title: MC IAM Manager API
  version: 0.6.2
paths:
...
  /api/permissions/mciam:
    post:
      consumes:
      - application/json
      description: Create a new permission with the specified information.
      operationId: createMciamPermission
      parameters:
      - description: Permission Info
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/model.MciamPermission'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.MciamPermission'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create new permission
      tags:
      - permissions
  /api/permissions/mciam/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a permission by its ID.
      operationId: deleteMciamPermission
      parameters:
      - description: Permission ID
        in: path
        name: permissionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete permission
      tags:
      - permissions
    put:
      consumes:
      - application/json
      description: Update the details of an existing permission.
      operationId: updateMciamPermission
      parameters:
      - description: Permission ID
        in: path
        name: permissionId
        required: true
        type: string
      - description: Permission Info
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/model.MciamPermission'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.MciamPermission'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update permission
      tags:
      - permissions
  /api/permissions/mciam/id/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve permission details by permission ID.
      operationId: getMciamPermissionByID
      parameters:
      - description: Permission ID
        in: path
        name: permissionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.MciamPermission'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get permission by ID
      tags:
      - permissions
  /api/permissions/mciam/list:
    post:
      consumes:
      - application/json
      description: Retrieve a list of all permissions.
      operationId: listMciamPermissions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.MciamPermission'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List all permissions
      tags:
      - permissions
...
  /api/roles/{roleType}/{roleId}/mciam-permissions:
    get:
      consumes:
      - application/json
      description: 특정 역할의 MC-IAM 권한 ID 목록을 조회합니다.
      operationId: getRoleMciamPermissions
      parameters:
      - description: 역할 타입 ('platform' or 'workspace')
        in: path
        name: roleType
        required: true
        type: string
      - description: 역할 ID
        in: path
        name: roleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 권한 ID 목록
          schema:
            items:
              type: string
            type: array
      summary: 역할의 MC-IAM 권한 목록 조회 - Renamed
      tags:
      - roles
      - mciam-permissions
  /api/roles/{roleType}/{roleId}/mciam-permissions/{permissionId}:
    delete:
      consumes:
      - application/json
      description: 역할에서 MC-IAM 권한을 제거합니다.
      operationId: removeMciamPermissionFromRole
      parameters:
      - description: 역할 타입 ('platform' or 'workspace')
        in: path
        name: roleType
        required: true
        type: string
      - description: 역할 ID
        in: path
        name: roleId
        required: true
        type: integer
      - description: MC-IAM 권한 ID
        in: path
        name: permissionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: 역할에서 MC-IAM 권한 제거 - Renamed
      tags:
      - roles
      - mciam-permissions
    post:
      consumes:
      - application/json
      description: 역할에 MC-IAM 권한을 할당합니다.
      operationId: assignMciamPermissionToRole
      parameters:
      - description: 역할 타입 ('platform' or 'workspace')
        in: path
        name: roleType
        required: true
        type: string
      - description: 역할 ID
        in: path
        name: roleId
        required: true
        type: integer
      - description: MC-IAM 권한 ID
        in: path
        name: permissionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: 역할에 MC-IAM 권한 할당 - Renamed
      tags:
      - roles
      - mciam-permissions
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"